+++ src/main/java/com/jaehyun/reservation/admin/store/service/StoreService.java
@@ -1,4 +1,4 @@
-package com.jaehyun.reservation.admin.smsservice.service;
+package com.jaehyun.reservation.admin.store.service;
 
 import com.jaehyun.reservation.admin.store.domain.dto.StoreReqDto;
 import com.jaehyun.reservation.admin.store.domain.dto.StoreResDto;
